Refactor table tests
